requirements.txt
django
djangorestframework
requests
gunicorn